refactor(advanced): 알림을 전역 Provider로 변경
- NotificationList를 NotificationProvider로 변경
- main.tsx에 전역 알림 Provider 추가
- App.tsx에서 알림 UI 제거하여 관심사 분리
- GlobalSVGProvider와 동일한 패턴 적용
