+++ src/advanced/app/App.tsx
@@ -1,6 +1,5 @@
 import { useState } from 'react';
 
-import { NotificationList } from '../entities/notification';
 import { AdminPage } from '../pages/admin';
 import { CartPage } from '../pages/cart';
 
@@ -9,8 +8,6 @@ export const App = () => {
 
   return (
     <div className='min-h-screen bg-gray-50'>
-      <NotificationList />
-
       {isAdmin ? (
         <AdminPage onChangeCartPage={() => setIsAdmin(false)} />
       ) : (
