+++ src/advanced/app/index.ts
@@ -1,4 +1,5 @@
 export { GlobalSVGProvider } from './providers/GlobalSVGProvider';
 export { JotaiProvider } from './providers/JotaiProvider';
+export { NotificationProvider } from './providers/NotificationProvider';
 
 export { App } from './App';
