+++ src/advanced/main.tsx
@@ -1,13 +1,14 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 
-import { App, GlobalSVGProvider, JotaiProvider } from './app';
+import { App, GlobalSVGProvider, JotaiProvider, NotificationProvider } from './app';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
     <JotaiProvider>
-      <App />
       <GlobalSVGProvider />
+      <NotificationProvider />
+      <App />
     </JotaiProvider>
   </React.StrictMode>
 );
